+++ deployment/terraform/cdn.tf
@@ -86,6 +86,7 @@ resource "aws_cloudfront_distribution" "cdn" {
   logging_config {
     include_cookies = false
     bucket          = "${aws_s3_bucket.logs.id}.s3.amazonaws.com"
+    prefix          = "CDN"
   }
 
   restrictions {